Changing Button Text
Added a referable string to the parent 'Ability' class and had each ability then Reset() it to its own description. Testing seems to indicate that it is working. One small error occurred when the Double Tap skill seemed to have lost the string value for its label for no apparent reason, but that was easily corrected via the inspector. tldr the buttons now change and hopefully this won't break anything. Will probably use a similar method to attach the descriptions so that access is easy.