Overworld Edit
Our Overworld looks a lot more like a desk now, with some stand-in textures just to get the point across. Now all it needs is some scattered pens, and maybe a stapler. But that's low priority and our art team is overworked as it is. Kudos to them for accomplishing as much as they have!